refactor(baselines): extract schema registry to dedicated module
