+++ baselines/maintainability.json
@@ -58,6 +58,7 @@
   ".agents/scripts/lib/audit-suite/workflow-loader.js": 115.677,
   ".agents/scripts/lib/auto-refresh-baselines.js": 87.316,
   ".agents/scripts/lib/baseline-loader.js": 105.28,
+  ".agents/scripts/lib/baseline-schema-registry.js": 103.244,
   ".agents/scripts/lib/baseline-snapshot.js": 78.303,
   ".agents/scripts/lib/bootstrap/baselines-layout-migration.js": 88.081,
   ".agents/scripts/lib/bootstrap/branch-protection.js": 94.023,
@@ -89,7 +90,7 @@
   ".agents/scripts/lib/close-validation/projections/maintainability.js": 96.56,
   ".agents/scripts/lib/config-gates-schema.js": 55.198,
   ".agents/scripts/lib/config-resolver.js": 84.277,
-  ".agents/scripts/lib/config-schema-shared.js": 100.936,
+  ".agents/scripts/lib/config-schema-shared.js": 138.317,
   ".agents/scripts/lib/config-schema.js": 171,
   ".agents/scripts/lib/config-settings-schema.js": 48.292,
   ".agents/scripts/lib/config/baselines.js": 105.01,
